# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Barefoot Nomad APIS",
        "description": "Testing endpoints"
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/user/confirmation/{email}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "confirm registrations",
                "description": "Confirmation of registrations",
                "operationId": "useremail",
                "parameters": [
                    {
                        "in": "path",
                        "name": "email",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userUpdate"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "confirmed"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "userUpdate": {
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
    }
}